agent/resume_processor_agent/requirements.txt
strands-agents>=0.2.0
strands-agents-tools>=0.2.0
boto3>=1.34.0
httpx>=0.27.0
bedrock-agentcore==0.1.7


